Split MILP code into separate file
